requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.5
django-admin-rangefilter==0.11.1
django-allauth==0.57.0
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
magic-filter==1.0.11
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
XlsxWriter==3.1.9
yarl==1.9.2
